Implement ForceHTTP10 Directive and SSLUncleanShutdown Directive
ForceHTTP10 will force the downgrading of HTTP/1.1 connections to 1.0 connections
based on the UserAgent string.  Similar to BrowserMatch and SetEnvIf from Apache.
However, this allows forcing non-ssl connections, as well as SSL connections (noHTTPS11)
and do it based on the browser regexes, so we can be more specific than .*MSIE.*

SSLUncleanShutdown matches against the user agent, and if it matches, does the
mod_ssl unclean shutdown behavior.  Compliant behavior is now the default.