BackendCookie implementation
Each Backend has a BackendKey which uniquely identifies the backend.
When BackendCookies are enabled, pound will inject a cookie into the response that will tag the browser to the specified backend.

In this way session affinity is handled in pound, without any need for session support on the backend servers, and sessions will persist across pound restarts because they're stored on the browser side.