write ipv4 addresses as ipv4 addresses, not ipv6 mapped addresses Based on a patch from the pound mailing list originally by Bussi Andrea <bussi@mfn.unipmn.it>
-----
If an IPv4 connection is made to an IPv6 socket then the local and
remote network addresses will be represented as IPv4-mapped addresses.
We try to print the IPv4 addresses in the usual format.

From inet_ntop man page:

BUGS
       AF_INET6 converts IPv4-mapped IPv6 addresses into an IPv6 format.

So this patch can be considered as a workaround for an inet_ntop bug.