Apply Frank Schmirler's RPC-over-HTTP patch
the attached patch fixes a problem with the RPC-over-HTTP code not working with NTLM/Kerberos authentication. A request must only be treated as an RPC if the content length is between 128kB and 2GB.

With the patch it is no longer necessary to enforce Basic auth for OutlookAnyhwere and the patch is the missing piece to support TerminalServicesGateway/RemoteDesktopGateway.